+++ src/light.c
@@ -1,6 +1,6 @@
 #include "rt.h"
 
-t_light	light_new(t_vec position, float intensity)
+t_light	light_new(t_vec position, double intensity)
 {
 	t_light light;
