+++ src/vector_utils.c
@@ -1,14 +1,14 @@
 #include "rt.h"
 
-float 	vec_length(t_vec v)
+double 	vec_length(t_vec v)
 {
-	return (sqrtf(vec_dot(v, v)));
+	return (sqrt(vec_dot(v, v)));
 }
 
 t_vec	vec_normalize(t_vec v)
 {
 	t_vec	new_v;
-	float	length;
+	double	length;
 
 	length = vec_length(v);
 	new_v.x = v.x / length;
@@ -17,7 +17,7 @@ t_vec	vec_normalize(t_vec v)
 	return (new_v);
 }
 
-float	vec_dot(t_vec v1, t_vec v2)
+double	vec_dot(t_vec v1, t_vec v2)
 {
 	return (v1.x * v2.x +
 			v1.y * v2.y +
